config/requirements.txt
# ==============================================
# RAG RCA Chatbot - Requirements
# Python 3.12+ recommended
# ==============================================

# === Core Framework ===
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# === LangChain Stack ===
langchain>=0.1.20
langchain-community>=0.0.38
langchain-core>=0.1.53
langchain-text-splitters>=0.0.2

# === Vector Store ===
chromadb>=0.4.22
langchain-chroma>=0.1.0

# === Embeddings & ML ===
sentence-transformers>=2.2.2
torch>=2.0.0
transformers>=4.36.0

# === PDF Processing ===
pypdf>=4.0.0
pymupdf>=1.23.0
pdfplumber>=0.10.0
pdfminer.six>=20221105

# === Hybrid Search (BM25) ===
rank-bm25>=0.2.2

# === Utilities ===
numpy>=1.24.0
tqdm>=4.66.0
requests>=2.31.0
truststore>=0.8.0
httpx>=0.27.0


# ===gui
streamlit
python-multipart